refactor: [8] 주소 로직 개선
