+++ src/lib/schema/profile.schema.ts
@@ -6,10 +6,13 @@ export const profileSchema = z.object({
     .min(1, { message: '이름을 입력해주세요.' })
     .max(6, { message: '이름은 최대 6자입니다.' })
     .regex(/^[가-힣]+$/, { message: '이름은 한글만 입력 가능합니다.' }),
-  address: z
-    .string()
-    .min(1, { message: '주소를 입력해주세요.' })
-    .max(100, { message: '주소는 100자 이내로 입력해주세요.' }),
+  address: z.object({
+    address: z.string(),
+    zonecode: z.string(),
+    sido: z.string(),
+    sigungu: z.string(),
+    bname: z.string(),
+  }),
   tags: z.array(z.string().min(1).max(15)).min(1).max(5),
 });
 export type ProfileType = z.infer<typeof profileSchema>;
@@ -23,10 +26,13 @@ export const stepOneSchema = z.object({
 });
 
 export const stepTwoSchema = z.object({
-  address: z
-    .string()
-    .min(1, { message: '주소를 입력해주세요.' })
-    .max(100, { message: '주소는 100자 이내로 입력해주세요.' }),
+  address: z.object({
+    address: z.string(),
+    zonecode: z.string(),
+    sido: z.string(),
+    sigungu: z.string(),
+    bname: z.string(),
+  }),
 });
 
 export const stepThreeSchema = z.object({
